Create main.cpp
This is the main.cpp for simulating performances of different initialization techniques (IniT) on EAs. Several int variables are used to control:
which EA and IniT to use
whether P or RA is used
whether best NP or fixed NP is used.
Also, the total FE can be controlled by FEperD. The function, number of P or RA, total runs can all be adjusted.